requirements.txt
transformers==4.52.3
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
requests>=2.28.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6